+++ src/main/java/com/umc/networkingService/domain/todoList/mapper/TodoListMapper.java
@@ -2,12 +2,13 @@ package com.umc.networkingService.domain.todoList.mapper;
 
 import com.umc.networkingService.domain.member.entity.Member;
 import com.umc.networkingService.domain.todoList.dto.request.TodoListCreateRequest;
-import com.umc.networkingService.domain.todoList.dto.request.TodoListUpdateRequest;
+import com.umc.networkingService.domain.todoList.dto.response.TodoListGetResponse;
 import com.umc.networkingService.domain.todoList.entity.ToDoList;
 import org.springframework.stereotype.Component;
 
 @Component
-public class TodoListCreateMapper {
+public class TodoListMapper {
+
     public ToDoList createTodoListToTodoList(Member member, TodoListCreateRequest request){
         return ToDoList.builder()
                 .writer(member)
@@ -16,4 +17,14 @@ public class TodoListCreateMapper {
                 .isCompleted(false)
                 .build();
     }
+
+    public TodoListGetResponse showTodoListToTodoList(ToDoList toDoList){
+        return TodoListGetResponse.builder()
+                .todoListId(toDoList.getId())
+                .title(toDoList.getTitle())
+                .deadline(toDoList.getDeadline())
+                .isCompleted(toDoList.isCompleted())
+                .build();
+    }
+
 }
